Separate tests  into its own file
